Made changes to test examples
