+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/LineGraph/BasicLineGraphs/ShapesHidden.test.jsx
@@ -3,12 +3,10 @@ import LineGraph from '@cerner/terra-graphs/lib/components/Line/LineGraph';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
 import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/dashedLineData';
+import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/simpleLineData';
 
-/*
-Please refer to the documentation below to see the graphConfig and data objects
-*/
-const graphConfig = utils.deepClone(lineDefault('#dashedLine'));
+const graphConfig = utils.deepClone(lineDefault('#shapesHidden'));
+graphConfig.showShapes = false;
 const dataset = utils.deepClone(data);
 
-export default () => <LineGraph graphID="dashedLine" graphConfig={graphConfig} dataset={dataset} />;
+export default () => <LineGraph graphID="shapesHidden" graphConfig={graphConfig} dataset={dataset} />;
