+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/LineGraph/BasicLineGraphs/SimpleLine.test.jsx
@@ -1,15 +1,17 @@
 import React from 'react';
 import LineGraph from '@cerner/terra-graphs/lib/components/Line/LineGraph';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import '@cerner/terra-graphs/lib/components/Graph.module.scss';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
 import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
 import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/simpleLineData';
 
-/*
-Please refer to the documentation below to see the graphConfig and data objects
-*/
-const graphConfig = utils.deepClone(lineDefault('#shapesHidden'));
-graphConfig.showShapes = false;
+const graphConfig = utils.deepClone(lineDefault('#simpleLinegraph'));
 const dataset = utils.deepClone(data);
 
-export default () => <LineGraph graphID="shapesHidden" graphConfig={graphConfig} dataset={dataset} />;
+export default () => (
+  <React.Fragment>
+    <div id="tooltip" className="initial-tooltip" />
+    <LineGraph graphID="simpleLinegraph" graphConfig={graphConfig} dataset={dataset} />
+  </React.Fragment>
+);
