+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/LineGraph/BasicLineGraphs/TimeseriesWithEventline.test.jsx
@@ -6,16 +6,28 @@ import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
 import lineTimeseries from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineTimeseries';
 import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/timeseriesData';
 
-/*
-Please refer to the documentation below to see the graphConfig and data objects
-*/
-const graphConfig = utils.deepClone(lineTimeseries('#simpleSpline'));
-const contentData = utils.deepClone(data);
-contentData[0].type = Carbon.helpers.LINE_TYPE.SPLINE;
+const graphConfig = utils.deepClone(lineTimeseries('#timeseriesEventLine'));
+graphConfig.eventline = [
+  {
+    color: Carbon.helpers.COLORS.GREY,
+    style: {
+      strokeDashArray: '4,4',
+    },
+    value: new Date(2016, 0, 1, 8).toISOString(),
+  },
+  {
+    color: Carbon.helpers.COLORS.BLACK,
+    style: {
+      strokeDashArray: '2,2',
+    },
+    value: new Date(2016, 0, 1, 12).toISOString(),
+  },
+];
+const dataset = utils.deepClone(data);
 
 export default () => (
   <React.Fragment>
     <div id="tooltip" className="initial-tooltip" />
-    <LineGraph graphID="simpleSpline" graphConfig={graphConfig} dataset={contentData} />
+    <LineGraph graphID="timeseriesEventLine" graphConfig={graphConfig} dataset={dataset} />
   </React.Fragment>
 );
