+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/LineGraph/Criticality/SimpleLine.test.jsx
@@ -3,25 +3,17 @@ import LineGraph from '@cerner/terra-graphs/lib/components/Line/LineGraph';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
 import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/defaultMultiLineData';
+import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/simpleLineData';
 
-/*
-Please refer to the documentation below to see the graphConfig and data objects
-*/
-const graphConfig = utils.deepClone(lineDefault('#criticalityMultiLine'));
+const graphConfig = utils.deepClone(lineDefault('#criticalitySimpleLine', 'LINE_TIMESERIES'));
 
 const contentData = utils.deepClone(data);
 contentData[0].values[0].isCritical = true;
-contentData[0].values[5].isCritical = true;
 contentData[0].values[10].isCritical = true;
-
-contentData[1].values[6].isCritical = true;
-contentData[1].values[11].isCritical = true;
 
 export default () => (
   <React.Fragment>
     <div id="tooltip" className="initial-tooltip" />
-    <LineGraph graphID="criticalityMultiLine" graphConfig={graphConfig} dataset={contentData} />
+    <LineGraph graphID="criticalitySimpleLine" graphConfig={graphConfig} dataset={contentData} />
   </React.Fragment>
 );
-
