+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/LineGraph/MultiLine/Simple.test.jsx
@@ -3,19 +3,15 @@ import LineGraph from '@cerner/terra-graphs/lib/components/Line/LineGraph';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
 import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/disableCalibrationData';
+import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/multiLineData';
 
-/*
-Please refer to the documentation below to see the graphConfig and data objects
-*/
-const graphConfig = utils.deepClone(lineDefault('#disableCalibration'));
-graphConfig.allowCalibration = false;
+const graphConfig = utils.deepClone(lineDefault('#multiLine'));
+const timeoutArray = [0, 750, 750 * 2, 750 * 3, 750 * 4, 750 * 5, 750 * 6];
 const dataset = utils.deepClone(data);
 
 export default () => (
   <React.Fragment>
     <div id="tooltip" className="initial-tooltip" />
-    <LineGraph graphID="disableCalibration" graphConfig={graphConfig} dataset={dataset} />
+    <LineGraph graphID="multiLine" graphConfig={graphConfig} dataset={dataset} timeout={timeoutArray} />
   </React.Fragment>
 );
-
