+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/LineGraph/NonContiguousLine/Simple.test.jsx
@@ -3,18 +3,14 @@ import LineGraph from '@cerner/terra-graphs/lib/components/Line/LineGraph';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
 import lineTimesries from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineTimeseries';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/valueRegionData';
+import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/nonContiguousLineData';
 
-/*
-Please refer to the documentation below to see the graphConfig and data objects
-*/
-const graphConfig = utils.deepClone(lineTimesries('#valueRegion'));
+const graphConfig = utils.deepClone(lineTimesries('#simpleLinegraph'));
 const dataset = utils.deepClone(data);
 
 export default () => (
   <React.Fragment>
     <div id="tooltip" className="initial-tooltip" />
-    <LineGraph graphID="valueRegion" graphConfig={graphConfig} dataset={dataset} />
+    <LineGraph graphID="simpleLinegraph" graphConfig={graphConfig} dataset={dataset} />
   </React.Fragment>
 );
-
