+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/LineGraph/Regions/Line/NoLowerBound.test.jsx
@@ -3,26 +3,19 @@ import LineGraph from '@cerner/terra-graphs/lib/components/Line/LineGraph';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
 import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/multiRegionData';
+import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/simpleLineData';
 
-/*
-Please refer to the documentation below to see the graphConfig and data objects
-*/
-const graphConfig = utils.deepClone(lineDefault('#regionLine'));
+const graphConfig = utils.deepClone(lineDefault('#noLowerBound'));
 const contentData = utils.deepClone(data);
-
 contentData[0].regions = [
   {
-    start: 10,
     end: 10,
-    color: '#bcbfc0',
   },
 ];
 
 export default () => (
   <React.Fragment>
     <div id="tooltip" className="initial-tooltip" />
-    <LineGraph graphID="regionLine" graphConfig={graphConfig} dataset={contentData} />
+    <LineGraph graphID="noLowerBound" graphConfig={graphConfig} dataset={contentData} />
   </React.Fragment>
 );
-
