+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/LineGraph/Regions/Line/NoUpperBound.test.jsx
@@ -3,18 +3,19 @@ import LineGraph from '@cerner/terra-graphs/lib/components/Line/LineGraph';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
 import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/multiLineData';
+import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/simpleLineData';
 
-/*
-Please refer to the documentation below to see the graphConfig and data objects
-*/
-const graphConfig = utils.deepClone(lineDefault('#multiLine'));
-const dataset = utils.deepClone(data);
-const timeout = [0, 750, 750 * 2, 750 * 3, 750 * 4, 750 * 5, 750 * 6];
+const graphConfig = utils.deepClone(lineDefault('#noUpperBound'));
+const contentData = utils.deepClone(data);
+contentData[0].regions = [
+  {
+    start: 2,
+  },
+];
 
 export default () => (
   <React.Fragment>
     <div id="tooltip" className="initial-tooltip" />
-    <LineGraph graphID="multiLine" graphConfig={graphConfig} dataset={dataset} timeout={timeout} />
+    <LineGraph graphID="noUpperBound" graphConfig={graphConfig} dataset={contentData} />
   </React.Fragment>
 );
