+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/LineGraph/Regions/Line/RegionsLine.test.jsx
@@ -3,21 +3,22 @@ import LineGraph from '@cerner/terra-graphs/lib/components/Line/LineGraph';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
 import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/simpleLineData';
+import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/multiRegionData';
 
-/*
-Please refer to the documentation below to see the graphConfig and data objects
-*/
-const graphConfig = utils.deepClone(lineDefault('#criticalitySimpleLine', 'LINE_TIMESERIES'));
-
+const graphConfig = utils.deepClone(lineDefault('#regionLine'));
 const contentData = utils.deepClone(data);
-contentData[0].values[0].isCritical = true;
-contentData[0].values[10].isCritical = true;
+
+contentData[0].regions = [
+  {
+    start: 10,
+    end: 10,
+    color: '#bcbfc0',
+  },
+];
 
 export default () => (
   <React.Fragment>
     <div id="tooltip" className="initial-tooltip" />
-    <LineGraph graphID="criticalitySimpleLine" graphConfig={graphConfig} dataset={contentData} />
+    <LineGraph graphID="regionLine" graphConfig={graphConfig} dataset={contentData} />
   </React.Fragment>
 );
-
