+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/LineGraph/Regions/MultiLine/DatasetRegionIdentical.test.jsx
@@ -5,14 +5,19 @@ import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
 import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
 import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/defaultMultiLineData';
 
-/*
-Please refer to the documentation below to see the graphConfig and data objects
-*/
-const graphConfig = utils.deepClone(lineDefault('#multipleline'));
+const graphConfig = utils.deepClone(lineDefault('#datasetRegionIdentical'));
 const contentData = utils.deepClone(data);
 contentData[0].regions = [
   {
     start: 2,
+    end: 14,
+  },
+];
+
+contentData[1].regions = [
+  {
+    start: 2,
+    end: 14,
   },
 ];
 
@@ -26,7 +31,6 @@ contentData[2].regions = [
 export default () => (
   <React.Fragment>
     <div id="tooltip" className="initial-tooltip" />
-    <LineGraph graphID="multipleline" graphConfig={graphConfig} dataset={contentData} />
+    <LineGraph graphID="datasetRegionIdentical" graphConfig={graphConfig} dataset={contentData} />
   </React.Fragment>
 );
-
