+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/LineGraph/Regions/MultiLine/MultiLine.test.jsx
@@ -3,23 +3,26 @@ import LineGraph from '@cerner/terra-graphs/lib/components/Line/LineGraph';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
 import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/simpleLineData';
+import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/defaultMultiLineData';
 
-/*
-Please refer to the documentation below to see the graphConfig and data objects
-*/
-const graphConfig = utils.deepClone(lineDefault('#noLowerBound'));
+const graphConfig = utils.deepClone(lineDefault('#multipleline'));
 const contentData = utils.deepClone(data);
 contentData[0].regions = [
   {
-    end: 10,
+    start: 2,
+  },
+];
+
+contentData[2].regions = [
+  {
+    start: 2,
+    end: 14,
   },
 ];
 
 export default () => (
   <React.Fragment>
     <div id="tooltip" className="initial-tooltip" />
-    <LineGraph graphID="noLowerBound" graphConfig={graphConfig} dataset={contentData} />
+    <LineGraph graphID="multipleline" graphConfig={graphConfig} dataset={contentData} />
   </React.Fragment>
 );
-
