+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/LineGraph/Regions/Spline/ValueRegion.test.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
+import LineGraph from '@cerner/terra-graphs/lib/components/Line/LineGraph';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
+import lineTimesries from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineTimeseries';
+import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/valueRegionData';
+
+const graphConfig = utils.deepClone(lineTimesries('#simpleSpline'));
+const contentData = utils.deepClone(data);
+contentData.type = Carbon.helpers.LINE_TYPE.SPLINE;
+
+export default () => (
+  <React.Fragment>
+    <div id="tooltip" className="initial-tooltip" />
+    <LineGraph graphID="simpleSpline" graphConfig={graphConfig} dataset={contentData} />
+  </React.Fragment>
+);
