+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/LineGraph/Spline/SimpleSpline.test.jsx
@@ -3,13 +3,10 @@ import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
 import LineGraph from '@cerner/terra-graphs/lib/components/Line/LineGraph';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
-import lineTimesries from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineTimeseries';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/valueRegionData';
+import lineTimeseries from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineTimeseries';
+import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/timeseriesData';
 
-/*
-Please refer to the documentation below to see the graphConfig and data objects
-*/
-const graphConfig = utils.deepClone(lineTimesries('#simpleSpline'));
+const graphConfig = utils.deepClone(lineTimeseries('#simpleSpline'));
 const contentData = utils.deepClone(data);
 contentData[0].type = Carbon.helpers.LINE_TYPE.SPLINE;
